Resolved a few TODO's.
maxReviewers is now a class wide field.
Cleaned up logging. The only info logging is maxReviewers and added
reviewers. All other logging is set to "debug".
addReviewers now takes a map instead of a set. The key is the account to
be added, and the value is the reason the reiewer is being added.

Change-Id: I4b4bf8148f75c5766e82fbf826f8729a442f4b50