Prevent review advice from overwriting other plugins' output
Instead of overwriting the HTML of the change_plugins element with
the review advice, put the review advice in a separate div and then
append that to the change_plugins element.

Strip down the string returned from the REST API to only include
the actual result without the surrounding div element(s).

Use a StringBuilder rather than string concatenation to build the
result.

Also fix the apply method to return a String rather than Object,
and replace multiple exception throws with RestApiException which
is the base exception of all of them.

Change-Id: I4e1b548e22baff75720c775398ec95165e381d5b